Refactor bot structure, improve entry point clarity, and update messaging templates
