+++ src/aura_telegram_bot/__init__.py
@@ -1,0 +1,1 @@
+"""Aura Telegram Bot: An intelligent assistant for your smart home."""
